requirements.txt
agate==1.7.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==14.0
croniter==2.0.1
cryptography==42.0.2
dagster==1.6.3
dagster-cloud==1.6.3
dagster-cloud-cli==1.6.3
dagster-graphql==1.6.3
dagster-pipes==1.6.3
dagster-webserver==1.6.3
dbt-core==1.7.7
dbt-extractor==0.5.1
dbt-postgres==1.7.7
dbt-semantic-interfaces==0.4.3
dlt==0.4.2
docstring-parser==0.15
e==1.4.5
fsspec==2024.2.0
gitdb==4.0.11
github3.py==4.0.1
GitPython==3.1.41
giturlparse==0.12.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-health-checking==1.60.1
h11==0.14.0
hexbytes==1.0.0
httptools==0.6.1
humanfriendly==10.0
humanize==4.9.0
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.3
jsonpath-ng==1.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
leather==0.3.4
Logbook==1.5.3
makefun==1.15.2
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mashumaro==3.12
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.2.0
msgpack==1.0.7
multidict==6.0.5
networkx==3.2.1
orjson==3.9.13
packaging==23.2
parsedatetime==2.6
pathspec==0.11.2
pathvalidate==3.2.0
pendulum==3.0.0
pex==2.1.163
pluggy==1.4.0
ply==3.11
prompt-toolkit==3.0.36
protobuf==4.25.2
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.3
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
requirements-parser==0.5.0
rich==13.7.0
rpds-py==0.17.1
semver==3.0.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.37.0
structlog==24.1.0
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
time-machine==2.13.0
tomli==2.0.1
tomlkit==0.12.3
toposort==1.10
tqdm==4.66.1
typer==0.9.0
types-setuptools==69.0.0.20240125
typing_extensions==4.9.0
tzdata==2023.4
universal-pathlib==0.1.4
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.27.0.post1
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.17.0
pyspark==3.5.2
kafka-python==2.0.2
django==4.2.15
django-jazzmin==3.0.0
